Test update

+++ src/Backend/Base/Config/Ent/OrgConfig.cs
@@ -1,6 +1,7 @@
 ﻿
 /// <summary>
 /// Configurations are for organisation
+/// Note this is not a database entity
 /// Created: June 2025
 /// [*Licence*]
 /// Author: John Stewart
@@ -10,7 +11,7 @@ namespace Backend.Base.Config.Ent
 {
     public class OrgConfig
     {
-        public long orgNr { get; set; }
+        public int orgNr { get; set; }
         public string LangCodeDefault { get; set; }
         public bool IsLangCodeEditable { get; set; } 
         public List<LanguageConfig> Languages { get; set; }
